requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-Migrate==4.0.5
python-dotenv==1.0.0
Werkzeug==2.3.7
Jinja2==3.1.2
WTForms==3.0.1
Flask-WTF==1.1.1
email-validator==2.0.0